delete multiple works now and is on home page also
